requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cryptography==41.0.1
dbus-python==1.2.16
defusedxml==0.7.1
distro-info===0.23ubuntu1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Fiona==1.8.20
GDAL==3.0.4
httplib2==0.22.0
idna==2.8
munch==3.0.0
numpy==1.24.3
oauthlib==3.2.2
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyGObject==3.36.0
PyJWT==2.7.0
pyparsing==3.0.9
python-apt==2.0.1+ubuntu0.20.4.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
six==1.14.0
sqlparse==0.4.4
typing-extensions==4.6.3
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
drf-yasg
django-rest-swagger